Moving asTaggableItem and TagButtons component to bodiless/filtering
