+++ packages/bodiless-filtering/src/FilterByGroup/asFilterableByGroup.tsx
@@ -13,10 +13,10 @@
  */
 
 import { withSidecarNodes } from '@bodiless/core';
-import { asTaggableItem } from '@bodiless/components';
 import { asToken } from '@bodiless/fclasses';
 import { withTagProps } from './FilterByGroupContext';
 import withFilterByTags from './withFilterByTags';
+import asTaggableItem from '../Taggable/asTaggableItem';
 
 const asFilterableByGroup = (nodeKey: string = '_tags') => asToken(
   withSidecarNodes(
